+++ Modules/DeviceInterface/devicerecord.cpp
@@ -1,16 +1,17 @@
 #include "devicerecord.h"
 
+
 DeviceRecord::record_t DeviceRecord::record{};
 
 
 uint16_t DeviceRecord::rcrd_id::ID;
 bool* DeviceRecord::rcrd_id::updated;
-void DeviceRecord::rcrd_id::set(uint64_t value, unsigned)
+void DeviceRecord::rcrd_id::set(uint32_t value, unsigned)
 {
     record.id = value;
 }
 
-uint64_t DeviceRecord::rcrd_id::get(unsigned)
+uint32_t DeviceRecord::rcrd_id::get(unsigned)
 {
     return record.id;
 }
@@ -18,12 +19,12 @@ uint64_t DeviceRecord::rcrd_id::get(unsigned)
 
 uint16_t DeviceRecord::time::ID;
 bool* DeviceRecord::time::updated;
-void DeviceRecord::time::set(uint64_t value, unsigned)
+void DeviceRecord::time::set(uint32_t value, unsigned)
 {
     record.time = value;
 }
 
-uint64_t DeviceRecord::time::get(unsigned)
+uint32_t DeviceRecord::time::get(unsigned)
 {
     return record.time;
 }
@@ -31,12 +32,12 @@ uint64_t DeviceRecord::time::get(unsigned)
 
 uint16_t DeviceRecord::MODBUS1_ID::ID;
 bool* DeviceRecord::MODBUS1_ID::updated;
-void DeviceRecord::MODBUS1_ID::set(uint64_t value, unsigned index)
+void DeviceRecord::MODBUS1_ID::set(uint8_t value, unsigned index)
 {
     record.mb1_sens[index].ID = value;
 }
 
-uint64_t DeviceRecord::MODBUS1_ID::get(unsigned index)
+uint8_t DeviceRecord::MODBUS1_ID::get(unsigned index)
 {
     return record.mb1_sens[index].ID;
 }
@@ -44,12 +45,12 @@ uint64_t DeviceRecord::MODBUS1_ID::get(unsigned index)
 
 uint16_t DeviceRecord::MODBUS1_value::ID;
 bool* DeviceRecord::MODBUS1_value::updated;
-void DeviceRecord::MODBUS1_value::set(uint64_t value, unsigned index)
+void DeviceRecord::MODBUS1_value::set(int16_t value, unsigned index)
 {
     record.mb1_sens[index].value = value;
 }
 
-uint64_t DeviceRecord::MODBUS1_value::get(unsigned index)
+int16_t DeviceRecord::MODBUS1_value::get(unsigned index)
 {
     return record.mb1_sens[index].value;
 }
@@ -70,12 +71,12 @@ uint64_t DeviceRecord::_1WIRE_ADDR::get(unsigned index)
 
 uint16_t DeviceRecord::_1WIRE_value::ID;
 bool* DeviceRecord::_1WIRE_value::updated;
-void DeviceRecord::_1WIRE_value::set(uint64_t value, unsigned index)
+void DeviceRecord::_1WIRE_value::set(int16_t value, unsigned index)
 {
     record.ow_sens[index].value = value;
 }
 
-uint64_t DeviceRecord::_1WIRE_value::get(unsigned index)
+int16_t DeviceRecord::_1WIRE_value::get(unsigned index)
 {
     return record.ow_sens[index].value;
 }
